requirements.txt
fastapi
uvicorn
python-multipart
faster-whisper
google-generativeai
